extract useParticipantSubmission hook
