+++ src/pages/iou/request/step/IOURequestStepParticipants.tsx
@@ -1,66 +1,25 @@
 import {useIsFocused} from '@react-navigation/core';
-import React, {useCallback, useEffect, useMemo, useRef} from 'react';
-import type {OnyxEntry} from 'react-native-onyx';
+import React, {useEffect} from 'react';
 import FormHelpMessage from '@components/FormHelpMessage';
-import useCurrentUserPersonalDetails from '@hooks/useCurrentUserPersonalDetails';
 import useLocalize from '@hooks/useLocalize';
-import useMappedPolicies from '@hooks/useMappedPolicies';
 import useOnyx from '@hooks/useOnyx';
-import useOptimisticDraftTransactions from '@hooks/useOptimisticDraftTransactions';
-import usePolicyForMovingExpenses from '@hooks/usePolicyForMovingExpenses';
+import useParticipantSubmission from '@hooks/useParticipantSubmission';
 import useThemeStyles from '@hooks/useThemeStyles';
-import useTransactionsByID from '@hooks/useTransactionsByID';
-import {setTransactionReport} from '@libs/actions/Transaction';
-import {READ_COMMANDS} from '@libs/API/types';
-import {isMobileSafari as isMobileSafariBrowser} from '@libs/Browser';
-import DistanceRequestUtils from '@libs/DistanceRequestUtils';
-import getPlatform from '@libs/getPlatform';
-import HttpUtils from '@libs/HttpUtils';
 import {isMovingTransactionFromTrackExpense as isMovingTransactionFromTrackExpenseIOUUtils, navigateToStartMoneyRequestStep} from '@libs/IOUUtils';
 import Navigation from '@libs/Navigation/Navigation';
-import {isPaidGroupPolicy} from '@libs/PolicyUtils';
-import {findSelfDMReportID, generateReportID, isInvoiceRoomWithID} from '@libs/ReportUtils';
-import {shouldRestrictUserBillableActions} from '@libs/SubscriptionUtils';
 import {endSpan} from '@libs/telemetry/activeSpans';
 import {getRequestType, hasRoute, isCorporateCardTransaction, isDistanceRequest, isPerDiemRequest, isTimeRequest as isTimeRequestUtil} from '@libs/TransactionUtils';
 import MoneyRequestParticipantsSelector from '@pages/iou/request/MoneyRequestParticipantsSelector';
-import {
-    resetDraftTransactionsCustomUnit,
-    setCustomUnitRateID,
-    setMoneyRequestCategory,
-    setMoneyRequestParticipants,
-    setMoneyRequestParticipantsFromReport,
-    setMoneyRequestTag,
-} from '@userActions/IOU';
 import {navigateToStartStepIfScanFileCannotBeRead} from '@userActions/IOU/Receipt';
-import {setSplitShares} from '@userActions/IOU/Split';
-import {createDraftWorkspace, generateDefaultWorkspaceName} from '@userActions/Policy/Policy';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
-import ROUTES from '@src/ROUTES';
 import type SCREENS from '@src/SCREENS';
-import {lastWorkspaceNumberSelector} from '@src/selectors/Policy';
-import type {Policy} from '@src/types/onyx';
-import type {Participant} from '@src/types/onyx/IOU';
 import getEmptyArray from '@src/types/utils/getEmptyArray';
-import KeyboardUtils from '@src/utils/keyboard';
 import StepScreenWrapper from './StepScreenWrapper';
 import type {WithFullTransactionOrNotFoundProps} from './withFullTransactionOrNotFound';
 import withFullTransactionOrNotFound from './withFullTransactionOrNotFound';
 import type {WithWritableReportOrNotFoundProps} from './withWritableReportOrNotFound';
 import withWritableReportOrNotFound from './withWritableReportOrNotFound';
-
-const policyMapper = (policy: OnyxEntry<Policy>): OnyxEntry<Policy> =>
-    policy && {
-        id: policy.id,
-        name: policy.name,
-        type: policy.type,
-        role: policy.role,
-        owner: policy.owner,
-        outputCurrency: policy.outputCurrency,
-        isPolicyExpenseChatEnabled: policy.isPolicyExpenseChatEnabled,
-        customUnits: policy.customUnits,
-    };
 
 type IOURequestStepParticipantsProps = WithWritableReportOrNotFoundProps<typeof SCREENS.MONEY_REQUEST.STEP_PARTICIPANTS> &
     WithFullTransactionOrNotFoundProps<typeof SCREENS.MONEY_REQUEST.STEP_PARTICIPANTS>;
@@ -75,73 +34,46 @@ function IOURequestStepParticipants({
     const {translate} = useLocalize();
     const styles = useThemeStyles();
     const isFocused = useIsFocused();
-    const {policyForMovingExpenses} = usePolicyForMovingExpenses();
     const [skipConfirmation] = useOnyx(`${ONYXKEYS.COLLECTION.SKIP_CONFIRMATION}${initialTransactionID}`);
-    const [allPolicies] = useMappedPolicies(policyMapper);
 
-    const [lastSelectedDistanceRates] = useOnyx(ONYXKEYS.NVP_LAST_SELECTED_DISTANCE_RATES);
-    const [draftTransactions] = useOptimisticDraftTransactions(initialTransaction);
-    // Depend on transactions.length to avoid updating transactionIDs when only the transaction details change
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    const transactionIDs = useMemo(() => draftTransactions?.map((transaction) => transaction.transactionID), [draftTransactions.length]);
-    const [transactions] = useTransactionsByID(transactionIDs);
-
-    // We need to set selectedReportID if user has navigated back from confirmation page and navigates to confirmation page with already selected participant
-    const selectedReportID = useRef<string>(participants?.length === 1 ? (participants.at(0)?.reportID ?? reportID) : reportID);
-    const numberOfParticipants = useRef(participants?.length ?? 0);
     const iouRequestType = getRequestType(initialTransaction);
     const isSplitRequest = iouType === CONST.IOU.TYPE.SPLIT;
     const isMovingTransactionFromTrackExpense = isMovingTransactionFromTrackExpenseIOUUtils(action);
-    const headerTitle = useMemo(() => {
-        if (action === CONST.IOU.ACTION.CATEGORIZE) {
-            return translate('iou.categorize');
-        }
-        if (action === CONST.IOU.ACTION.SHARE) {
-            return translate('iou.share');
-        }
-        if (isSplitRequest) {
-            return translate('iou.splitExpense');
-        }
-        if (iouType === CONST.IOU.TYPE.PAY) {
-            return translate('iou.paySomeone');
-        }
-        if (iouType === CONST.IOU.TYPE.INVOICE) {
-            return translate('workspace.invoices.sendInvoice');
-        }
-        return translate('iou.chooseRecipient');
-    }, [iouType, translate, isSplitRequest, action]);
-
-    const selfDMReportID = findSelfDMReportID();
-    const [selfDMReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${selfDMReportID}`);
-    const [activePolicyID] = useOnyx(ONYXKEYS.NVP_ACTIVE_POLICY_ID);
-    const [activePolicy] = useOnyx(`${ONYXKEYS.COLLECTION.POLICY}${activePolicyID}`);
-    const [introSelected] = useOnyx(ONYXKEYS.NVP_INTRO_SELECTED);
-    const [userBillingGracePeriodEnds] = useOnyx(ONYXKEYS.COLLECTION.SHARED_NVP_PRIVATE_USER_BILLING_GRACE_PERIOD_END);
-    const [ownerBillingGracePeriodEnd] = useOnyx(ONYXKEYS.NVP_PRIVATE_OWNER_BILLING_GRACE_PERIOD_END);
-    const [amountOwed] = useOnyx(ONYXKEYS.NVP_PRIVATE_AMOUNT_OWED);
-
-    const currentUserPersonalDetails = useCurrentUserPersonalDetails();
-
-    const isActivePolicyRequest =
-        iouType === CONST.IOU.TYPE.CREATE &&
-        isPaidGroupPolicy(activePolicy) &&
-        activePolicy?.isPolicyExpenseChatEnabled &&
-        !shouldRestrictUserBillableActions(activePolicy.id, ownerBillingGracePeriodEnd, userBillingGracePeriodEnds, amountOwed);
-
-    const isAndroidNative = getPlatform() === CONST.PLATFORM.ANDROID;
-    const isMobileSafari = isMobileSafariBrowser();
     const isPerDiem = isPerDiemRequest(initialTransaction);
     const isTime = isTimeRequestUtil(initialTransaction);
     const isCorporateCard = isCorporateCardTransaction(initialTransaction);
+
+    let headerTitle = translate('iou.chooseRecipient');
+    if (action === CONST.IOU.ACTION.CATEGORIZE) {
+        headerTitle = translate('iou.categorize');
+    } else if (action === CONST.IOU.ACTION.SHARE) {
+        headerTitle = translate('iou.share');
+    } else if (isSplitRequest) {
+        headerTitle = translate('iou.splitExpense');
+    } else if (iouType === CONST.IOU.TYPE.PAY) {
+        headerTitle = translate('iou.paySomeone');
+    } else if (iouType === CONST.IOU.TYPE.INVOICE) {
+        headerTitle = translate('workspace.invoices.sendInvoice');
+    }
+
+    const {addParticipant, goToNextStep} = useParticipantSubmission({
+        reportID,
+        initialTransactionID,
+        initialTransaction,
+        participants,
+        iouType,
+        action,
+        backTo,
+        isSplitRequest,
+        isMovingTransactionFromTrackExpense,
+        isFocused,
+    });
 
     useEffect(() => {
         endSpan(CONST.TELEMETRY.SPAN_OPEN_CREATE_EXPENSE);
     }, []);
 
     // When the component mounts, if there is a receipt, see if the image can be read from the disk. If not, redirect the user to the starting step of the flow.
-    // This is because until the expense is saved, the receipt file is only stored in the browsers memory as a blob:// and if the browser is refreshed, then
-    // the image ceases to exist. The best way for the user to recover from this is to start over from the start of the expense process.
-    // skip this in case user is moving the transaction as the receipt path will be valid in that case
     useEffect(() => {
         if (isMovingTransactionFromTrackExpense) {
             return;
@@ -161,292 +93,21 @@ function IOURequestStepParticipants({
         isMovingTransactionFromTrackExpense,
     ]);
 
-    // When the step opens, reset the draft transaction's custom unit if moved from Track Expense.
-    // This resets the custom unit to the p2p rate when the destination workspace changes,
-    // because we want to first check if the p2p rate exists on the workspace.
-    // If it doesn't exist - we'll show an error message to force the user to choose a valid rate from the workspace.
-    useEffect(() => {
-        if (!isMovingTransactionFromTrackExpense || !isFocused || !transactions || transactions?.length === 0) {
-            return;
-        }
-
-        for (const transaction of transactions) {
-            resetDraftTransactionsCustomUnit(transaction);
-        }
-    }, [isFocused, isMovingTransactionFromTrackExpense, transactions]);
-
-    const waitForKeyboardDismiss = useCallback(
-        (callback: () => void) => {
-            if (isAndroidNative || isMobileSafari) {
-                KeyboardUtils.dismiss().then(() => {
-                    callback();
-                });
-            } else {
-                callback();
-            }
-        },
-        [isAndroidNative, isMobileSafari],
-    );
-
-    const trackExpense = useCallback(() => {
-        // If coming from the combined submit/track flow and the user proceeds to just track the expense,
-        // we will use the track IOU type in the confirmation flow.
-        if (!selfDMReportID) {
-            return;
-        }
-
-        for (const transaction of draftTransactions) {
-            const rateID = DistanceRequestUtils.getCustomUnitRateID({
-                reportID: selfDMReportID,
-                isTrackDistanceExpense: isDistanceRequest(transaction),
-                policy: policyForMovingExpenses,
-                isPolicyExpenseChat: false,
-                lastSelectedDistanceRates,
-            });
-            setCustomUnitRateID(transaction.transactionID, rateID, transaction, policyForMovingExpenses);
-            const shouldSetParticipantAutoAssignment = iouType === CONST.IOU.TYPE.CREATE;
-            setMoneyRequestParticipantsFromReport(
-                transaction.transactionID,
-                selfDMReport,
-                currentUserPersonalDetails.accountID,
-                shouldSetParticipantAutoAssignment ? isActivePolicyRequest : false,
-            );
-            setTransactionReport(transaction.transactionID, {reportID: CONST.REPORT.UNREPORTED_REPORT_ID}, true);
-        }
-        const iouConfirmationPageRoute = ROUTES.MONEY_REQUEST_STEP_CONFIRMATION.getRoute(action, CONST.IOU.TYPE.TRACK, initialTransactionID, selfDMReportID);
-        waitForKeyboardDismiss(() => {
-            // If the backTo parameter is set, we should navigate back to the confirmation screen that is already on the stack.
-            Navigation.setNavigationActionToMicrotaskQueue(() => {
-                if (backTo) {
-                    // We don't want to compare params because we just changed the participants.
-                    Navigation.goBack(iouConfirmationPageRoute, {compareParams: false});
-                } else {
-                    // We wrap navigation in setNavigationActionToMicrotaskQueue so that data loading in Onyx and navigation do not occur simultaneously, which resets the amount to 0.
-                    // More information can be found here: https://github.com/Expensify/App/issues/73728
-                    Navigation.navigate(iouConfirmationPageRoute);
-                }
-            });
-        });
-    }, [
-        selfDMReportID,
-        draftTransactions,
-        action,
-        initialTransactionID,
-        waitForKeyboardDismiss,
-        iouType,
-        selfDMReport,
-        currentUserPersonalDetails.accountID,
-        isActivePolicyRequest,
-        backTo,
-        policyForMovingExpenses,
-        lastSelectedDistanceRates,
-    ]);
-
-    const addParticipant = useCallback(
-        (val: Participant[]) => {
-            HttpUtils.cancelPendingRequests(READ_COMMANDS.SEARCH_FOR_REPORTS);
-
-            const firstParticipant = val.at(0);
-
-            if (firstParticipant?.isSelfDM && !isSplitRequest) {
-                trackExpense();
-                return;
-            }
-
-            const firstParticipantReportID = val.at(0)?.reportID;
-            const isPolicyExpenseChat = !!firstParticipant?.isPolicyExpenseChat;
-            const policy = isPolicyExpenseChat && firstParticipant?.policyID ? allPolicies?.[`${ONYXKEYS.COLLECTION.POLICY}${firstParticipant.policyID}`] : undefined;
-            const isInvoice = iouType === CONST.IOU.TYPE.INVOICE;
-            numberOfParticipants.current = val.length;
-
-            // Use transactions array if available, otherwise use initialTransactionID directly
-            // This handles the case where initialTransaction hasn't loaded yet but we still need to set participants
-            if (draftTransactions.length > 0) {
-                for (const transaction of draftTransactions) {
-                    setMoneyRequestParticipants(transaction.transactionID, val);
-                }
-            } else {
-                // Fallback to using initialTransactionID directly when transaction object isn't loaded yet
-                setMoneyRequestParticipants(initialTransactionID, val);
-            }
-
-            if (!isMovingTransactionFromTrackExpense || !isPolicyExpenseChat) {
-                // If not moving the transaction from track expense to a workspace, select the default rate automatically.
-                // Otherwise, keep the original rate and let the user manually change it to the one they want from the workspace.
-                const rateID = DistanceRequestUtils.getCustomUnitRateID({reportID: firstParticipantReportID, isPolicyExpenseChat, policy, lastSelectedDistanceRates});
-
-                if (draftTransactions.length > 0) {
-                    for (const transaction of draftTransactions) {
-                        setCustomUnitRateID(transaction.transactionID, rateID, transaction, policy);
-                    }
-                } else {
-                    // Fallback to using initialTransactionID directly
-                    setCustomUnitRateID(initialTransactionID, rateID, undefined, policy);
-                }
-            }
-
-            // When multiple valid participants are selected, the reportID is generated at the end of the confirmation step.
-            // So we are resetting selectedReportID ref to the reportID coming from params.
-            // For invoices, a valid participant must have a login.
-
-            const hasOneValidParticipant = iouType === CONST.IOU.TYPE.INVOICE && selectedReportID.current !== reportID ? val.filter((item) => !!item.login).length !== 1 : val.length !== 1;
-
-            if (hasOneValidParticipant && !isInvoice) {
-                selectedReportID.current = reportID;
-                return;
-            }
-
-            // When a participant is selected, the reportID needs to be saved because that's the reportID that will be used in the confirmation step.
-            // We use || to be sure that if the first participant doesn't have a reportID, we generate a new one.
-            if (isInvoice) {
-                selectedReportID.current = firstParticipantReportID && isInvoiceRoomWithID(firstParticipantReportID) ? firstParticipantReportID : generateReportID();
-            } else {
-                // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-                selectedReportID.current = firstParticipantReportID || generateReportID();
-            }
-        },
-        [isSplitRequest, allPolicies, iouType, draftTransactions, isMovingTransactionFromTrackExpense, reportID, trackExpense, initialTransactionID, lastSelectedDistanceRates],
-    );
-
-    const goToNextStep = useCallback(
-        (_value?: string, _participants?: Participant[]) => {
-            const isCategorizing = action === CONST.IOU.ACTION.CATEGORIZE;
-            const isShareAction = action === CONST.IOU.ACTION.SHARE;
-
-            const isPolicyExpenseChat = participants?.some((participant) => participant.isPolicyExpenseChat);
-            if (iouType === CONST.IOU.TYPE.SPLIT && !isPolicyExpenseChat && initialTransaction?.amount && initialTransaction?.currency) {
-                const participantAccountIDs = participants?.map((participant) => participant.accountID) as number[];
-                setSplitShares(initialTransaction, initialTransaction.amount, initialTransaction.currency, participantAccountIDs);
-            }
-
-            const newReportID = selectedReportID.current;
-            const shouldUpdateTransactionReportID = participants?.at(0)?.reportID !== newReportID;
-            const transactionReportID = newReportID === selfDMReportID ? CONST.REPORT.UNREPORTED_REPORT_ID : newReportID;
-            // TODO: probably should also change participants here for selectedParticipants.current, but out of scope of this PR
-            for (const transaction of draftTransactions) {
-                const tag = isMovingTransactionFromTrackExpense && transaction?.tag ? transaction?.tag : '';
-                setMoneyRequestTag(transaction.transactionID, tag);
-                const firstParticipant = _participants?.at(0);
-                const policy = isPolicyExpenseChat && firstParticipant?.policyID ? allPolicies?.[`${ONYXKEYS.COLLECTION.POLICY}${firstParticipant.policyID}`] : undefined;
-                const policyDistance = Object.values(policy?.customUnits ?? {}).find((customUnit) => customUnit.name === CONST.CUSTOM_UNITS.NAME_DISTANCE);
-                const defaultCategory = isDistanceRequest(transaction) && policyDistance?.defaultCategory ? policyDistance?.defaultCategory : '';
-                const category = isMovingTransactionFromTrackExpense ? (transaction?.category ?? '') : defaultCategory;
-                setMoneyRequestCategory(transaction.transactionID, category, isMovingTransactionFromTrackExpense ? policyForMovingExpenses : undefined, isMovingTransactionFromTrackExpense);
-                if (shouldUpdateTransactionReportID) {
-                    setTransactionReport(transaction.transactionID, {reportID: transactionReportID}, true);
-                }
-            }
-            if ((isCategorizing || isShareAction) && numberOfParticipants.current === 0) {
-                const email = currentUserPersonalDetails.email ?? '';
-                const lastWorkspaceNumber = lastWorkspaceNumberSelector(allPolicies, email);
-                const {expenseChatReportID, policyID, policyName} = createDraftWorkspace(
-                    introSelected,
-                    generateDefaultWorkspaceName(email, lastWorkspaceNumber, translate),
-                    currentUserPersonalDetails.accountID,
-                    email,
-                );
-                for (const transaction of draftTransactions) {
-                    setMoneyRequestParticipants(transaction.transactionID, [
-                        {
-                            selected: true,
-                            accountID: 0,
-                            isPolicyExpenseChat: true,
-                            reportID: expenseChatReportID,
-                            policyID,
-                            searchText: policyName,
-                        },
-                    ]);
-                }
-                Navigation.setNavigationActionToMicrotaskQueue(() => {
-                    if (isCategorizing) {
-                        Navigation.navigate(ROUTES.MONEY_REQUEST_STEP_CATEGORY.getRoute(action, CONST.IOU.TYPE.SUBMIT, initialTransactionID, expenseChatReportID));
-                    } else {
-                        Navigation.navigate(ROUTES.MONEY_REQUEST_STEP_CONFIRMATION.getRoute(action, CONST.IOU.TYPE.SUBMIT, initialTransactionID, expenseChatReportID, undefined, true));
-                    }
-                });
-                return;
-            }
-
-            const iouConfirmationPageRoute = ROUTES.MONEY_REQUEST_STEP_CONFIRMATION.getRoute(
-                action,
-                iouType === CONST.IOU.TYPE.CREATE || iouType === CONST.IOU.TYPE.TRACK ? CONST.IOU.TYPE.SUBMIT : iouType,
-                initialTransactionID,
-                newReportID,
-                undefined,
-                undefined,
-                action === CONST.IOU.ACTION.SHARE ? Navigation.getActiveRoute() : undefined,
-            );
-
-            const route = isCategorizing
-                ? ROUTES.MONEY_REQUEST_STEP_CATEGORY.getRoute(action, iouType, initialTransactionID, selectedReportID.current || reportID, iouConfirmationPageRoute)
-                : iouConfirmationPageRoute;
-
-            waitForKeyboardDismiss(() => {
-                // If the backTo parameter is set, we should navigate back to the confirmation screen that is already on the stack.
-                // We wrap navigation in setNavigationActionToMicrotaskQueue so that data loading in Onyx and navigation do not occur simultaneously, which resets the amount to 0.
-                // More information can be found here: https://github.com/Expensify/App/issues/73728
-                Navigation.setNavigationActionToMicrotaskQueue(() => {
-                    if (backTo) {
-                        // We don't want to compare params because we just changed the participants.
-                        Navigation.goBack(route, {compareParams: false});
-                    } else {
-                        Navigation.navigate(route);
-                    }
-                });
-            });
-        },
-        [
-            action,
-            participants,
-            iouType,
-            initialTransaction,
-            selfDMReportID,
-            initialTransactionID,
-            reportID,
-            waitForKeyboardDismiss,
-            draftTransactions,
-            isMovingTransactionFromTrackExpense,
-            allPolicies,
-            policyForMovingExpenses,
-            currentUserPersonalDetails.email,
-            currentUserPersonalDetails.accountID,
-            introSelected,
-            translate,
-            backTo,
-        ],
-    );
-
-    const navigateBack = useCallback(() => {
+    const navigateBack = () => {
         if (backTo) {
             Navigation.goBack(backTo);
             return;
         }
 
-        // Change iouType param to enable negative values
         const shouldForceIOUType =
             action === CONST.IOU.ACTION.CREATE && iouType === CONST.IOU.TYPE.SUBMIT && (iouRequestType === CONST.IOU.REQUEST_TYPE.MANUAL || iouRequestType === CONST.IOU.REQUEST_TYPE.SCAN);
         const iouTypeValue = shouldForceIOUType ? CONST.IOU.TYPE.CREATE : iouType;
 
         navigateToStartMoneyRequestStep(iouRequestType, iouTypeValue, initialTransactionID, reportID, action);
-    }, [backTo, iouRequestType, iouType, initialTransactionID, reportID, action]);
+    };
 
-    useEffect(() => {
-        const isCategorizing = action === CONST.IOU.ACTION.CATEGORIZE;
-        const isShareAction = action === CONST.IOU.ACTION.SHARE;
-        if (isFocused && (isCategorizing || isShareAction)) {
-            for (const transaction of draftTransactions) {
-                setMoneyRequestParticipants(transaction.transactionID, []);
-            }
-            numberOfParticipants.current = 0;
-        }
-        // We don't want to clear out participants every time the transactions change
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [isFocused, action]);
-
-    const isWorkspacesOnly = useMemo(() => {
+    const getIsWorkspacesOnly = () => {
         if (isDistanceRequest(initialTransaction)) {
-            // For distance requests, only restrict to workspaces if a route exists and the distance is 0
-            // If no route exists yet, the distance hasn't been calculated and we should allow P2P
             if (!hasRoute(initialTransaction, true)) {
                 return false;
             }
@@ -458,7 +119,8 @@ function IOURequestStepParticipants({
         }
 
         return initialTransaction?.amount !== undefined && initialTransaction?.amount !== null && initialTransaction?.amount <= 0;
-    }, [initialTransaction, iouRequestType]);
+    };
+    const isWorkspacesOnly = getIsWorkspacesOnly();
 
     return (
         <StepScreenWrapper
